turned buildGrid function into a component
