+++ src/pages/home-page.js
@@ -9,6 +9,7 @@ import DropdownSortMenu from '../components/dropdown-sort-menu';
 import ToggleOtherTagsButton from '../components/toggle-other-tags-button.js';
 import LabelExampleIcon from '../components/tags.js';
 import ViewOptionsToggle from '../components/view-options-toggle.js';
+import BuildGrid from '../components/build-grid.js';
 import { LoadingContext } from '../App.js';
 import '../styles/home-page.css';
 
@@ -74,31 +75,31 @@ const HomePage = (props) => {
     });
   };
 
-  // This function creates the grid of course review cards
-  const buildGrid = () => {
-    const sortedCourses = sortMostReviewed();
-    const gridArray = [];
-    const colSize = 3;
-    for (let i = 0; i < sortedCourses.length; i += colSize) {
-      const gridRow = sortedCourses.slice(i, i + colSize);
-      gridArray.push(gridRow);
-    }
-    return gridArray.map((row, index) => {
-      return (
-        <Grid.Row key={index}>
-          {row.map((course) => (
-            <Grid.Column key={course.id}>
-              <CourseReviewCard
-                code={course.courseCode}
-                name={course.title}
-                desc={course.description}
-                numReviews={course.reviews.length}
-              />
-            </Grid.Column>))}
-        </Grid.Row>
-      );
-    });
-  };
+  // // This function creates the grid of course review cards
+  // const buildGrid = () => {
+  //   const sortedCourses = sortMostReviewed();
+  //   const gridArray = [];
+  //   const colSize = 3;
+  //   for (let i = 0; i < sortedCourses.length; i += colSize) {
+  //     const gridRow = sortedCourses.slice(i, i + colSize);
+  //     gridArray.push(gridRow);
+  //   }
+  //   return gridArray.map((row, index) => {
+  //     return (
+  //       <Grid.Row key={index}>
+  //         {row.map((course) => (
+  //           <Grid.Column key={course.id}>
+  //             <CourseReviewCard
+  //               code={course.courseCode}
+  //               name={course.title}
+  //               desc={course.description}
+  //               numReviews={course.reviews.length}
+  //             />
+  //           </Grid.Column>))}
+  //       </Grid.Row>
+  //     );
+  //   });
+  // };
 
   const handleQueryChange = (e, { value }) => {
     setQuery(value);
@@ -185,7 +186,7 @@ const HomePage = (props) => {
 
       {/* Code, name and desc hardcoded for testing purposes */}
       <Grid columns={3}>
-        {buildGrid()}
+        <BuildGrid courses={courses}></BuildGrid>
       </Grid>
     </>
   );
